assignment-03/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
